requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
distro==1.9.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
snownlp==0.12.3
jiter==0.12.0
joblib==1.5.2
llvmlite==0.45.1
pymilvus[milvus_lite]==2.6.3
pymilvus[model]
numba==0.62.1
numpy==2.3.4
openai==2.7.2
orjson==3.11.4
pandas==2.3.3
protobuf==6.33.0
pyahocorasick==2.2.0
pydantic==2.12.4
pydantic_core==2.41.5
pymilvus==2.6.3
PyMySQL==1.1.2
pynndescent==0.5.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
